+++ airbox/plotters/met.py
@@ -5,6 +5,8 @@ import pandas as pd
 from airbox.dir import get_instr_dir
 from .base import BasePlotter
 
+from pandas.io.common import _NA_VALUES
+
 
 class MetPlotter(BasePlotter):
     name = 'met'
@@ -60,7 +62,9 @@ class MetPlotter(BasePlotter):
         :return: Pandas dataframe
         """
         fname = '{}-{:02}-{:02}Values.Txt'.format(d.year, d.month, d.day)
+
+        na_values = list(_NA_VALUES) + ['No answer', 'FC: 40']
         data = pd.read_csv(join(get_instr_dir('WeatherStation'), fname), parse_dates=[0], index_col=0,
-                           names=self.columns, header=None, skiprows=5, sep=';')
+                           names=self.columns, header=None, skiprows=5, sep=';', na_values=na_values)
 
         return data[data.index.date == d]