+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 
 - Read in daily Spectronus data and added to plots
 - Added subset command for extracting a data for a single instrument
+- Remove nan values from met data
 
 0.1.1
 -----